requirement.txt
airtest==1.1.1
certifi==2019.9.11
chardet==3.0.4
comtypes==1.1.7
facebook-wda==0.2.1
hrpc==1.0.8
idna==2.8
Jinja2==2.10.3
MarkupSafe==1.1.1
mss==4.0.3
numpy==1.17.3
opencv-contrib-python==3.4.2.17
opencv-python==4.1.1.26
Pillow==6.2.1
pocoui==1.0.78
pypiwin32==223
pywin32==225
pywinauto==0.6.3
requests==2.22.0
six==1.12.0
urllib3==1.25.6
websocket-client==0.48.0
